requirements.txt
Flask==3.0.0
openai==1.3.7
python-dotenv==1.0.0
Flask-WTF==1.2.1
gunicorn==21.2.0
